Update joke.py
This modified code finds the index of the first line that contains "![Funny Meme]" and replaces it with the new markdown. If no such line is found, it doesn't make any changes to the file. This way, it only changes the first line with "![Funny Meme]" while leaving the rest of the content intact.